Update Bixtral Parameters models.mdx
I'm not 100% sure of the math, but shouldnt it be closer to 140Bish? I guess it's the 8x22=176 but since Mixtral 8x7B actually displays 45B taking into consideration the shared weights I guess this is a mistake? There is the same mistake at the open weight models section too btw !